Move processors into a single location
