+++ app/src/test/java/gov/va/vro/service/MasAssessmentResultProcessorTest.java
@@ -6,7 +6,7 @@ import gov.va.vro.BaseIntegrationTest;
 import gov.va.vro.model.rrd.AbdEvidenceWithSummary;
 import gov.va.vro.model.rrd.mas.MasAutomatedClaimPayload;
 import gov.va.vro.persistence.repository.AssessmentResultRepository;
-import gov.va.vro.service.provider.services.MasAssessmentResultProcessor;
+import gov.va.vro.service.provider.camel.processor.MasAssessmentResultProcessor;
 import gov.va.vro.service.spi.db.SaveToDbService;
 import gov.va.vro.service.spi.model.Claim;
 import org.apache.camel.Exchange;
