+++ domain-rrd/rrd-workflows/src/main/java/gov/va/vro/service/provider/camel/PrimaryRoutes.java
@@ -3,8 +3,8 @@ package gov.va.vro.service.provider.camel;
 import gov.va.vro.camel.FunctionProcessor;
 import gov.va.vro.camel.RabbitMqCamelUtils;
 import gov.va.vro.camel.ToRabbitMqRouteHelper;
-import gov.va.vro.service.provider.services.AssessmentResultProcessor;
-import gov.va.vro.service.provider.services.EvidenceSummaryDocumentProcessor;
+import gov.va.vro.service.provider.camel.processor.AssessmentResultProcessor;
+import gov.va.vro.service.provider.camel.processor.EvidenceSummaryDocumentProcessor;
 import gov.va.vro.service.spi.db.SaveToDbService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
