+++ domain-rrd/rrd-workflows/src/main/java/gov/va/vro/service/provider/camel/processor/MasOrderExamProcessor.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider;
+package gov.va.vro.service.provider.camel.processor;
 
 import gov.va.vro.model.rrd.mas.MasAutomatedClaimPayload;
 import gov.va.vro.model.rrd.mas.MasOrderExamConditions;
