+++ domain-rrd/rrd-workflows/src/main/java/gov/va/vro/service/provider/camel/processor/ServiceLocationsExtractorProcessor.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.camel;
+package gov.va.vro.service.provider.camel.processor;
 
 import gov.va.vro.model.rrd.AbdEvidence;
 import gov.va.vro.model.rrd.HealthDataAssessment;
