+++ domain-rrd/rrd-workflows/src/main/java/gov/va/vro/service/provider/utils/DiagnosisLookup.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.services;
+package gov.va.vro.service.provider.utils;
 
 import java.util.Map;
